requirements.txt
# requirements.txt

# Core Application
fastapi
uvicorn[standard]
pydantic
python-dotenv
python-multipart

# AI & Workflow
langchain
langgraph
xgboost
shap
scikit-learn
imbalanced-learn
pandas
numpy
joblib
great-expectations

# Database & Storage
sqlalchemy
psycopg2-binary>=2.9
azure-storage-blob
httpx>=0.27




